Create a generic useEndowmentsType redux-toolkit hook V2
